+++ src/SerializerGenerator.php
@@ -314,7 +314,9 @@ final class SerializerGenerator
         array $stack,
     ): string {
         $index = '$index'.mb_strlen($arrayPath);
-        $resolvedModelPath = "{$index}Array";
+        $value = '$value'.mb_strlen($arrayPath);
+        $itemArrayPath = $arrayPath.'['.$index.']';
+
         $subType = $type->getSubType();
 
         switch ($subType) {
@@ -324,15 +326,15 @@ final class SerializerGenerator
                 return $this->templating->renderArrayAssign($arrayPath, $modelPath);
 
             case $subType instanceof PropertyTypeIterable:
-                $innerCode = $this->generateCodeForArray($subType, $apiVersion, $serializerGroups, $arrayPath.'['.$index.']', $resolvedModelPath.'['.$index.']', $stack);
+                $innerCode = $this->generateCodeForArray($subType, $apiVersion, $serializerGroups, $itemArrayPath, $value, $stack);
                 break;
 
             case $subType instanceof PropertyTypeEnum:
-                $innerCode = $this->generateCodeForFieldType($subType, $apiVersion, $serializerGroups, $arrayPath.'['.$index.']', $resolvedModelPath.'['.$index.']', $stack);
+                $innerCode = $this->generateCodeForFieldType($subType, $apiVersion, $serializerGroups, $itemArrayPath, $value, $stack);
                 break;
 
             case $subType instanceof PropertyTypeClass:
-                $innerCode = $this->generateCodeForClass($subType->getClassMetadata(), $apiVersion, $serializerGroups, $arrayPath.'['.$index.']', $resolvedModelPath.'['.$index.']', $stack);
+                $innerCode = $this->generateCodeForClass($subType->getClassMetadata(), $apiVersion, $serializerGroups, $itemArrayPath, $value, $stack);
                 break;
 
             default:
@@ -348,10 +350,10 @@ final class SerializerGenerator
         }
 
         if ($type->isHashmap()) {
-            return $this->templating->renderLoopHashmap($arrayPath, $modelPath, $index, $innerCode);
+            return $this->templating->renderLoopHashmap($arrayPath, $modelPath, $index, $value, $innerCode);
         }
 
-        return $this->templating->renderLoopArray($arrayPath, $modelPath, $index, $innerCode);
+        return $this->templating->renderLoopArray($arrayPath, $modelPath, $index, $value, $innerCode);
     }
 
     /**